Move common search/test utils methods into the base package
