+++ courtfinder/apps/search/tests/test_search_statistics.py
@@ -4,14 +4,14 @@ from django.utils.timezone import now
 
 from courtfinder.test_utils import TestCaseWithData
 from search.models import SearchStatistic
-from search.tests.test_search import valid_response
+from . import postcode_valid
 
 
 class SearchStatisticsTestCase(TestCaseWithData):
 
     def test_latest_search(self):
         client = Client()
-        with valid_response('lookup_postcode'):
+        with postcode_valid('lookup_postcode'):
             self.assertIsNone(SearchStatistic.get_latest_search())
             response = client.get(reverse('search:results'), data={
                 'aol': 'crime',
@@ -25,7 +25,7 @@ class SearchStatisticsTestCase(TestCaseWithData):
     def test_pingdom_doesnt_count(self):
         client = Client(
             HTTP_USER_AGENT='Pingdom.com_bot_version_1.4_(http://www.pingdom.com/)')
-        with valid_response('lookup_postcode'):
+        with postcode_valid('lookup_postcode'):
             self.assertIsNone(SearchStatistic.get_latest_search())
             response = client.get(reverse('search:results'), data={
                 'aol': 'crime',
